refactor: public side events list separated components.
